Add Delete Method, fix bugs and polish

+++ servicos/ProdutoAPI/src/main/java/com/codefootmen/softwareengineer/api/repository/ProdutosRepository.java
@@ -3,5 +3,5 @@ package com.codefootmen.softwareengineer.api.repository;
 import com.codefootmen.softwareengineer.api.model.Produtos;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProdutosRepository extends JpaRepository<Produtos, Long> {
+public interface ProdutosRepository extends JpaRepository<Produtos, Integer> {
 }
